backend/requirements.txt
asgiref==3.7.2
azure-core==1.29.5
azure-storage-blob==12.19.0
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
dj-database-url==2.1.0
Django==4.2.6
django-cors-headers==4.3.0
django-on-heroku==1.1.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
google==3.0.0
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth==2.24.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud==0.34.0
google-cloud-bigquery==3.13.0
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-cloud-vision==3.4.5
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
oauth2client==4.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.2
Pillow==10.1.0
proto-plus==1.22.3
protobuf==4.25.1
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyDrive==1.3.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
whitenoise==6.6.0
